Delegate some stuff out of Starship
