+++ server/src/main/kotlin/net/stellarica/server/craft/CraftContainer.kt
@@ -1,6 +1,9 @@
 package net.stellarica.server.craft
 
-interface CraftContainer : Craft {
+import net.stellarica.common.coordinate.BlockPosition
+
+interface CraftContainer {
 	fun addSubCraft(craft: Craft)
 	fun removeSubCraft(craft: Craft)
+	fun subcraftsContain(block: BlockPosition): Boolean
 }
