+++ server/src/main/kotlin/net/stellarica/server/craft/Pilotable.kt
@@ -2,7 +2,7 @@ package net.stellarica.server.craft
 
 import org.bukkit.entity.Player
 
-interface Pilotable : Craft {
+interface Pilotable : Rideable {
 	val pilot: Player?
 
 	val isPiloted: Boolean
@@ -11,4 +11,7 @@ interface Pilotable : Craft {
 	fun pilot(pilot: Player)
 
 	fun unpilot()
+
+	// todo: I don't like this being part of the interface
+	fun movePassengers(craft: Craft, data: CraftTransformation)
 }
