+++ server/src/main/kotlin/net/stellarica/server/craft/SimpleCraftContainer.kt
@@ -1,0 +1,17 @@
+package net.stellarica.server.craft
+
+import net.stellarica.common.coordinate.BlockPosition
+
+class SimpleCraftContainer : CraftContainer {
+	override fun addSubCraft(craft: Craft) {
+		TODO("Not yet implemented")
+	}
+
+	override fun removeSubCraft(craft: Craft) {
+		TODO("Not yet implemented")
+	}
+
+	override fun subcraftsContain(block: BlockPosition): Boolean {
+		TODO()
+	}
+}
