+++ server/src/main/kotlin/net/stellarica/server/craft/starship/subsystem/Subsystem.kt
@@ -1,4 +1,3 @@
 package net.stellarica.server.craft.starship.subsystem
 
-interface Subsystem {
-}
+interface Subsystem
